[fix] : ignore new migration for test environment

+++ database/migrations/2021_04_28_182547_add_startdate_and_enddate_and_remove_deadline_from_projects_table.php
@@ -14,9 +14,13 @@ class AddStartdateAndEnddateAndRemoveDeadlineFromProjectsTable extends Migration
     public function up()
     {
         Schema::table('projects', function (Blueprint $table) {
-            $table->date('startdate')->after('max_member_count')->nullable();
-			$table->date('enddate')->after('startdate')->nullable();
-            $table->dropColumn('deadline');
+
+            if (app()->environment() !== 'testing') {
+                $table->date('startdate')->after('max_member_count')->nullable();
+			    $table->date('enddate')->after('startdate')->nullable();
+                $table->dropColumn('deadline');
+            }
+
         });
     }
 
@@ -28,9 +32,13 @@ class AddStartdateAndEnddateAndRemoveDeadlineFromProjectsTable extends Migration
     public function down()
     {
         Schema::table('projects', function (Blueprint $table) {
-            $table->dropColumn('startdate');
-            $table->dropColumn('enddate');
-            $table->date('deadline')->nullable();
+
+            if (app()->environment() !== 'testing') {
+                $table->dropColumn('startdate');
+                $table->dropColumn('enddate');
+                $table->date('deadline')->nullable();
+            }
+
         });
     }
 }
